requirements.txt
flask
flask-restful
mysql-connector-python


streamlit
pandas
requests
matplotlib
seaborn
